refactor(customer): customer 패키지 분리
