+++ src/main/java/woowacourse/shoppingcart/application/CustomerService.java
@@ -9,8 +9,8 @@ import woowacourse.shoppingcart.application.dto.CustomerPasswordUpdateServiceReq
 import woowacourse.shoppingcart.application.dto.CustomerProfileUpdateServiceRequest;
 import woowacourse.shoppingcart.application.dto.CustomerSaveServiceRequest;
 import woowacourse.shoppingcart.dao.CustomerDao;
-import woowacourse.shoppingcart.domain.Customer;
-import woowacourse.shoppingcart.domain.Password;
+import woowacourse.shoppingcart.domain.customer.Customer;
+import woowacourse.shoppingcart.domain.customer.Password;
 import woowacourse.shoppingcart.domain.PasswordConvertor;
 import woowacourse.shoppingcart.exception.DuplicatedEmailException;
 import woowacourse.shoppingcart.exception.InvalidCustomerException;
