+++ src/main/java/woowacourse/shoppingcart/domain/customer/Customer.java
@@ -1,4 +1,6 @@
-package woowacourse.shoppingcart.domain;
+package woowacourse.shoppingcart.domain.customer;
+
+import woowacourse.shoppingcart.domain.PasswordConvertor;
 
 public class Customer {
