+++ src/main/java/woowacourse/shoppingcart/dto/CustomerRegisterRequest.java
@@ -4,7 +4,7 @@ import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Size;
 import woowacourse.shoppingcart.application.dto.CustomerSaveServiceRequest;
-import woowacourse.shoppingcart.domain.Password;
+import woowacourse.shoppingcart.domain.customer.Password;
 
 public class CustomerRegisterRequest {
