+++ src/test/java/woowacourse/shoppingcart/application/CustomerServiceTest.java
@@ -23,8 +23,8 @@ import woowacourse.shoppingcart.application.dto.CustomerPasswordUpdateServiceReq
 import woowacourse.shoppingcart.application.dto.CustomerProfileUpdateServiceRequest;
 import woowacourse.shoppingcart.application.dto.CustomerSaveServiceRequest;
 import woowacourse.shoppingcart.dao.CustomerDao;
-import woowacourse.shoppingcart.domain.Customer;
-import woowacourse.shoppingcart.domain.Password;
+import woowacourse.shoppingcart.domain.customer.Customer;
+import woowacourse.shoppingcart.domain.customer.Password;
 import woowacourse.shoppingcart.domain.PasswordConvertor;
 import woowacourse.shoppingcart.exception.DuplicatedEmailException;
